Integration for player

See merge request compsci308_2019spring/voogasalad_nomergeconflicts!83

+++ src/Configurations/src/Configs/ArsenalConfig/WeaponBehaviors/PlaceableOnPath.java
@@ -22,7 +22,7 @@ public class PlaceableOnPath extends WeaponBehavior {
     }
 
     @Override
-    public void update(double ms) {
+    public void update(long ms) {
 
     }
